chrome-page-saver/pyproject.toml
[project]
name = "chrome-page-saver"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
]
